+++ src/main/kotlin/dc8/xCore/commands/essentials/GamemodeCmd.kt
@@ -1,10 +1,10 @@
 package dc8.xCore.commands.essentials
 
 import dc8.xCore.XCore
-import dc8.xCore.commands.Helper
-import dc8.xCore.commands.Helper.appendLine
-import dc8.xCore.commands.Helper.append
-import dc8.xCore.commands.Helper.hasAnyPermissionOf
+import dc8.xCore.commands.append
+import dc8.xCore.commands.appendLine
+import dc8.xCore.commands.fail
+import dc8.xCore.commands.suggestPlayers
 import io.papermc.paper.command.brigadier.BasicCommand
 import io.papermc.paper.command.brigadier.CommandSourceStack
 import net.kyori.adventure.text.Component
@@ -17,12 +17,131 @@ import org.bukkit.GameMode
 import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
 import java.util.logging.Level
-import dc8.xCore.permissions.Perms.GM as Perms
+import dc8.xCore.globals.Permissions.GM as Perms
+
+/**
+ * Represents the structured and validated parameters of the command.
+ * @property gameMode the gamemode to change to
+ * @property target the target player
+ */
+private data class GamemodeArgs(val gameMode: GameMode, val target: Player)
+
+class GamemodeCmd(private val xCore: XCore) : BasicCommand {
+    override fun execute(context: CommandSourceStack, rawArgs: Array<out String>) {
+        val args = validate(context, rawArgs) ?: return
+
+        if (args.target.gameMode == args.gameMode)
+            return
+
+        args.target.gameMode = args.gameMode
+
+        args.target.sendMessage(Responses.success(args.gameMode.name.lowercase()))
+
+        xCore.logger.log(
+            Level.FINER,
+            "${context.sender} changed gamemode of ${args.target.name} to ${args.gameMode.name.lowercase()}."
+        )
+    }
+
+    override fun canUse(sender: CommandSender): Boolean =
+        listOf(
+            Perms.SURVIVAL,
+            Perms.CREATIVE,
+            Perms.ADVENTURE,
+            Perms.SPECTATOR
+        ).any {
+            sender.hasPermission(it)
+        }
+
+    override fun suggest(context: CommandSourceStack, args: Array<out String>): List<String> {
+        return when {
+            args.size <= 1 -> getPermittedGamemodes(context.sender)
+            args.size == 2 && context.sender.hasPermission(Perms.OTHERS) ->
+                suggestPlayers(args[1], context.sender)
+
+            else -> emptyList()
+        }
+    }
+
+    /**
+     * Checks the permissions of the given [CommandSender] and builds a list of valid input
+     * suggestions as the gamemode parameter.
+     * @param sender the command sender.
+     * @return a list of valid [String] inputs based on permission limitations.
+     */
+    private fun getPermittedGamemodes(sender: CommandSender): List<String> {
+        return buildList(4) {
+            if (sender.hasPermission(Perms.SURVIVAL)) add("0")
+            if (sender.hasPermission(Perms.CREATIVE)) add("1")
+            if (sender.hasPermission(Perms.ADVENTURE)) add("2")
+            if (sender.hasPermission(Perms.SPECTATOR)) add("3")
+        }
+    }
+
+    /**
+     * Validates command arguments and maps them to [GamemodeArgs].
+     *
+     * @param context the command source context
+     * @param rawArgs the raw arguments passed to the command
+     * @return the parsed [GamemodeArgs], or `null` if validation fails
+     */
+    private fun validate(context: CommandSourceStack, rawArgs: Array<out String>): GamemodeArgs? {
+        val sender = context.sender
+
+        // valid number of args and valid gamemode?
+        val gamemodeInt = rawArgs.firstOrNull()?.toIntOrNull()
+        if (gamemodeInt == null || gamemodeInt !in 0..3)
+            return sender.fail(Responses.invalidSyntax)
+
+        // is the target valid?
+        if (rawArgs.size == 1 && context.executor !is Player)
+            return sender.fail(Responses.invalidTarget)
+        if (rawArgs.size == 2) {
+            if (!sender.hasPermission(Perms.OTHERS))
+                return sender.fail(Responses.permissionError)
+
+            if (Bukkit.getPlayer(rawArgs[1]) == null)
+                return sender.fail(Responses.invalidTarget)
+        }
+
+        // has permission for this gamemode?
+        val gamemode = mapGamemode(gamemodeInt)
+        if (!sender.hasPermission(Perms.BASE_MODE + gamemodeInt))
+            return sender.fail(Responses.permissionError)
+
+        val targetPlayer = when(rawArgs.size) {
+            2 -> Bukkit.getPlayer(rawArgs[1])!!
+            else -> context.executor as Player
+        }
+
+        return GamemodeArgs(gamemode, targetPlayer)
+    }
+
+    /**
+     * Maps from an [Int] to a [GameMode].
+     * @param gamemodeId the ID of the gamemode.
+     * @return the respective [GameMode].
+     * @throws IllegalArgumentException when the ID does not match a [GameMode].
+     */
+    private fun mapGamemode(gamemodeId: Int): GameMode {
+        return when (gamemodeId) {
+            0 -> GameMode.SURVIVAL
+            1 -> GameMode.CREATIVE
+            2 -> GameMode.ADVENTURE
+            3 -> GameMode.SPECTATOR
+            else -> throw IllegalArgumentException() // Should never happen
+        }
+    }
+}
 
 /**
  * Predefined error messages used by the command.
  */
-private object ErrorResponses {
+private object Responses {
+    fun success(gamemode: String): Component =
+        text("Gamemode changed to ", NamedTextColor.GREEN)
+            .append(gamemode, NamedTextColor.LIGHT_PURPLE)
+
     /**
      * The target is invalid.
      */
@@ -66,127 +185,4 @@ private object ErrorResponses {
     val permissionError = text(
         "You don't have permission to use this command (like this).", NamedTextColor.RED
     )
-}
-
-/**
- * Represents the structured and validated parameters of the command.
- * @property gameMode the gamemode to change to
- * @property target the target player
- */
-private class GamemodeArgs(val gameMode: GameMode, val target: Player)
-
-class GamemodeCmd(private val xCore: XCore) : BasicCommand {
-    override fun execute(context: CommandSourceStack, rawArgs: Array<out String>) {
-        val args = validate(context, rawArgs) ?: return
-
-        if (args.target.gameMode == args.gameMode) {
-            return
-        }
-
-        args.target.gameMode = args.gameMode
-
-        args.target.sendMessage(
-            text("Gamemode changed to ", NamedTextColor.GREEN)
-                .append(args.gameMode.name.lowercase(), NamedTextColor.LIGHT_PURPLE)
-        )
-
-        xCore.logger.log(
-            Level.FINER,
-            "${context.sender} changed gamemode of ${args.target.name} to ${args.gameMode.name.lowercase()}."
-        )
-    }
-
-    override fun canUse(sender: CommandSender): Boolean = sender.hasAnyPermissionOf(
-        Perms.SURVIVAL,
-        Perms.CREATIVE,
-        Perms.ADVENTURE,
-        Perms.SPECTATOR
-    )
-
-    override fun suggest(context: CommandSourceStack, args: Array<out String>): List<String> {
-        return when {
-            args.size <= 1 -> getPermittedGamemodes(context.sender)
-            args.size == 2 && context.sender.hasPermission(Perms.OTHERS) ->
-                Helper.suggestPlayers(args[1], context.sender)
-
-            else -> emptyList()
-        }
-    }
-
-    /**
-     * Checks the permissions of the given [CommandSender] and builds a list of valid input
-     * suggestions as the gamemode parameter.
-     * @param sender the command sender.
-     * @return a list of valid [String] inputs based on permission limitations.
-     */
-    private fun getPermittedGamemodes(sender: CommandSender): List<String> {
-        return buildList(4) {
-            if (sender.hasPermission(Perms.SURVIVAL)) add("0")
-            if (sender.hasPermission(Perms.CREATIVE)) add("1")
-            if (sender.hasPermission(Perms.ADVENTURE)) add("2")
-            if (sender.hasPermission(Perms.SPECTATOR)) add("3")
-        }
-    }
-
-    /**
-     * Validates command arguments and maps them to [GamemodeArgs].
-     *
-     * @param context the command source context
-     * @param rawArgs the raw arguments passed to the command
-     * @return the parsed [GamemodeArgs], or `null` if validation fails
-     */
-    private fun validate(context: CommandSourceStack, rawArgs: Array<out String>): GamemodeArgs? {
-        val sender = context.sender
-
-        fun fail(message: Component): Nothing? {
-            sender.sendMessage(message)
-            return null
-        }
-
-        // valid number of args and valid gamemode?
-        val gamemodeInt = rawArgs.firstOrNull()?.toIntOrNull()
-        if (gamemodeInt == null || gamemodeInt !in 0..3) {
-            return fail(ErrorResponses.invalidSyntax)
-        }
-
-        // is the target valid?
-        if (rawArgs.size == 1 && context.executor !is Player) {
-            return fail(ErrorResponses.invalidTarget)
-        }
-        if (rawArgs.size == 2) {
-            if (!sender.hasPermission(Perms.OTHERS)) {
-                return fail(ErrorResponses.permissionError)
-            }
-            if (Bukkit.getPlayer(rawArgs[1]) == null) {
-                return fail(ErrorResponses.invalidTarget)
-            }
-        }
-
-        // has permission for this gamemode?
-        val gamemode = mapGamemode(gamemodeInt)
-        if (!sender.hasPermission(Perms.COMMAND_BASE_MODE + gamemodeInt)) {
-            return fail(ErrorResponses.permissionError)
-        }
-
-        val targetPlayer = if (rawArgs.size == 2) Bukkit.getPlayer(rawArgs[1])!!
-        else context.executor as Player
-
-        return GamemodeArgs(gamemode, targetPlayer)
-    }
-
-    /**
-     * Maps from an [Int] to a [GameMode].
-     * @param gamemodeId the ID of the gamemode.
-     * @return the respective [GameMode].
-     * @throws IllegalArgumentException when the ID does not match a [GameMode].
-     */
-    private fun mapGamemode(gamemodeId: Int): GameMode {
-        return when (gamemodeId) {
-            0 -> GameMode.SURVIVAL
-            1 -> GameMode.CREATIVE
-            2 -> GameMode.ADVENTURE
-            3 -> GameMode.SPECTATOR
-            else -> throw IllegalArgumentException() // Should never happen
-        }
-    }
 }