+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # X-Core Changelog
-### 0.0.1
+### 1.0.0-alpha
+* [EDIT] Project structure
+* [EDIT] Permission naming convention
+* [NEW] Database Support
+* [NEW] `/xcore reload`-command
+* [NEW] `/home`-command
+* [NEW] `/homes`-command
+* [NEW] `/sethome`-command
+* [NEW] `/deletehome`-command
+* [NEW] `/locatehome`-command
+
+### 0.0.1-alpha
 * [EDIT] KDocs
 
 ### Initial Commit